time_sync: time_sync_last_epoch als RTC time_t (ESP32-C6 64-bit)
Ersetzt uint64_t durch RTC_DATA_ATTR time_t fuer semantisch korrekte Epoch-Speicherung und entfernt ueberfluessige Casts bei time(NULL) - time_sync_last_epoch.

+++ src/transfer_mqtt.cpp
@@ -248,7 +248,7 @@ transfer_status_t transfer_mqtt_send_data(const transfer_data_t* data) {
     }
 
     bool ntp_ok = sync_ntp_time();
-    int64_t ntp_status = ntp_ok ? (int64_t)time_sync_last_epoch : -1;
+    int64_t ntp_status = ntp_ok ? time_sync_last_epoch : -1;
 
     char timestamp_iso[40] = "";
     if (ntp_ok) {
